Refactor EGA (again) - implement CGA emulation
