+++ core/src/devices/ega/draw.rs
@@ -44,30 +44,29 @@ impl EGACard {
     #[inline]
     pub fn draw_solid_hchar_6bpp(&mut self, color: u8) {
         let frame_u64: &mut [u64] = bytemuck::cast_slice_mut(&mut *self.buf[self.back_buf]);
-
-        let attr_color = EGA_COLORS_6BPP_U64[(color & 0x3F) as usize];
-        frame_u64[self.rba >> 3] = attr_color;
+        frame_u64[self.rba >> 3] = EGA_COLORS_6BPP_U64[(color & 0x3F) as usize];
     }
 
+    #[inline]
     pub fn draw_debug_hchar_at(&mut self, addr: usize, color: u8) {
         let frame_u64: &mut [u64] = bytemuck::cast_slice_mut(&mut *self.buf[self.back_buf]);
         frame_u64[addr >> 3] = EGA_COLORS_U64[(color & 0x0F) as usize];
     }
 
-    /// Draw a character in lowres mode (16 pixels) using a single solid color.
+    /// Draw a character in low res mode (16 pixels) using a single solid color.
     /// Since all pixels are the same we can draw 64 bits at a time.
     #[inline]
     pub fn draw_solid_lchar(&mut self, color: u8) {
         let frame_u64: &mut [u64] = bytemuck::cast_slice_mut(&mut *self.buf[self.back_buf]);
-        frame_u64[self.rba >> 3] = EGA_COLORS_U64[(color & 0x0F) as usize];
-        frame_u64[(self.rba >> 3) + 1] = EGA_COLORS_U64[(color & 0x0F) as usize];
+        let attr_color = EGA_COLORS_U64[(color & 0x0F) as usize];
+        frame_u64[self.rba >> 3] = attr_color;
+        frame_u64[(self.rba >> 3) + 1] = attr_color;
     }
 
     #[inline]
     pub fn draw_solid_lchar_6bpp(&mut self, color: u8) {
         let frame_u64: &mut [u64] = bytemuck::cast_slice_mut(&mut *self.buf[self.back_buf]);
-
-        let attr_color = EGA_COLORS_6BPP_U64[self.attribute_palette_registers[(color & 0x0F) as usize].six as usize];
+        let attr_color = EGA_COLORS_6BPP_U64[self.ac.attribute_palette_registers[(color & 0x0F) as usize].six as usize];
         frame_u64[self.rba >> 3] = attr_color;
         frame_u64[(self.rba >> 3) + 1] = attr_color;
     }
@@ -75,17 +74,13 @@ impl EGACard {
     /// Draw an entire character row in high resolution text mode (8 pixels)
     pub fn draw_text_mode_hchar14(&mut self) {
         // Do cursor if visible, enabled and defined
-        if self.vma == self.crtc_cursor_address as usize
-            && self.cursor_status
-            && self.blink_state
-            && self.cursor_data[(self.vlc & 0x3F) as usize]
-        {
+        if self.crtc.status.cursor {
             self.draw_solid_hchar_6bpp(self.cur_fg);
         }
         else if self.mode_enable {
             let glyph_row: u64;
             // Get the u64 glyph row to draw for the current fg and bg colors and character row (vlc)
-            glyph_row = self.get_hchar_glyph14_row(self.cur_char as usize, self.vlc as usize);
+            glyph_row = self.get_hchar_glyph14_row(self.cur_char as usize, self.crtc.vlc() as usize);
 
             let frame_u64: &mut [u64] = bytemuck::cast_slice_mut(&mut *self.buf[self.back_buf]);
             frame_u64[self.rba >> 3] = glyph_row;
@@ -99,16 +94,13 @@ impl EGACard {
     /// Draw an entire character row in low resolution text mode (16 pixels)
     pub fn draw_text_mode_lchar14(&mut self) {
         // Do cursor if visible, enabled and defined
-        if self.vma == self.crtc_cursor_address as usize
-            && self.cursor_status
-            && self.blink_state
-            && self.cursor_data[(self.vlc & 0x3F) as usize]
-        {
+        if self.crtc.status.cursor {
             self.draw_solid_lchar_6bpp(self.cur_fg);
         }
         else if self.mode_enable {
             // Get the two u64 glyph row components to draw for the current fg and bg colors and character row (vlc)
-            let (glyph_row0, glyph_row1) = self.get_lchar_glyph14_rows(self.cur_char as usize, self.vlc as usize);
+            let (glyph_row0, glyph_row1) =
+                self.get_lchar_glyph14_rows(self.cur_char as usize, self.crtc.vlc() as usize);
 
             let frame_u64: &mut [u64] = bytemuck::cast_slice_mut(&mut *self.buf[self.back_buf]);
             frame_u64[self.rba >> 3] = glyph_row0;
@@ -127,7 +119,7 @@ impl EGACard {
 
         for i in 0..8 {
             let buf_i = i - self.pel_pan_latch as usize;
-            self.buf[self.back_buf][self.rba + buf_i] = self.chain_buf[(self.vma * 8 & 0x7FFFF) + i];
+            self.buf[self.back_buf][self.rba + buf_i] = self.sequencer.read_linear((self.vma * 8 & 0x7FFFF) + i);
         }
     }
 
@@ -139,8 +131,9 @@ impl EGACard {
         for i in 0..8 {
             let buf_i = i - self.pel_pan_latch as usize;
 
-            let attr_color =
-                self.attribute_palette_registers[(self.chain_buf[(self.vma * 8 & 0x7FFFF) + i] & 0x0F) as usize].six;
+            let attr_color = self.ac.attribute_palette_registers
+                [(self.sequencer.read_linear((self.vma * 8 & 0x7FFFF) + i) & 0x0F) as usize]
+                .six;
             self.buf[self.back_buf][self.rba + buf_i] = attr_color;
         }
     }
@@ -150,11 +143,10 @@ impl EGACard {
         //let deplaned_u64: &mut [u64] = bytemuck::cast_slice_mut(&mut *self.chain_buf);
         //frame_u64[self.rba >> 3] = deplaned_u64[(self.vma & 0xFFFFF) >> 3];
 
+        let ser = self.gc.serialize(&self.sequencer, self.vma);
         for i in 0..8 {
             let buf_i = (i * 2) - (self.pel_pan_latch * 2) as usize;
-            let attr_color = self.attribute_palette_registers
-                [(self.chain_buf[(self.vma * 8 & 0x7FFFF) + i] & 0x3F) as usize]
-                .four_to_six;
+            let attr_color = self.ac.attribute_palette_registers[(ser[i] & 0x3F) as usize].four_to_six;
             self.buf[self.back_buf][self.rba + buf_i] = attr_color;
             self.buf[self.back_buf][self.rba + buf_i + 1] = attr_color;
         }
@@ -168,8 +160,9 @@ impl EGACard {
         for i in 0..8 {
             let buf_i = (i * 2) - (self.pel_pan_latch * 2) as usize;
 
-            let attr_color =
-                self.attribute_palette_registers[(self.chain_buf[(self.vma * 8 & 0x7FFFF) + i] & 0x3F) as usize].six;
+            let attr_color = self.ac.attribute_palette_registers
+                [(self.sequencer.read_linear((self.vma * 8 & 0x7FFFF) + i) & 0x3F) as usize]
+                .six;
             self.buf[self.back_buf][self.rba + buf_i] = attr_color;
             self.buf[self.back_buf][self.rba + buf_i + 1] = attr_color;
         }
