+++ core/src/devices/ega/videocard.rs
@@ -49,8 +49,16 @@ impl VideoCard for EGACard {
         (false, false, false, false)
     }
 
-    fn set_video_option(&mut self, _opt: VideoOption) {
-        // No options implemented
+    fn set_video_option(&mut self, opt: VideoOption) {
+        match opt {
+            VideoOption::EnableSnow(state) => {
+                log::warn!("VideoOption::EnableSnow not supported for EGA");
+            }
+            VideoOption::DebugDraw(state) => {
+                log::debug!("VideoOption::DebugDraw set to: {}", state);
+                self.debug_draw = state;
+            }
+        }
     }
 
     fn get_video_type(&self) -> VideoType {
@@ -74,13 +82,13 @@ impl VideoCard for EGACard {
     }
 
     fn get_display_size(&self) -> (u32, u32) {
-        // EGA supports multiple fonts.
+        /*        // EGA supports multiple fonts.
 
         let font_w = EGA_FONTS[self.current_font].w;
         let _font_h = EGA_FONTS[self.current_font].h;
 
         // Clock divisor effectively doubles the CRTC register values
-        let _clock_divisor = match self.sequencer_clocking_mode.dot_clock() {
+        let _clock_divisor = match self.sequencer.clocking_mode.dot_clock() {
             DotClock::Native => 1,
             DotClock::HalfClock => 2,
         };
@@ -88,7 +96,13 @@ impl VideoCard for EGACard {
         //let width = (self.crtc_horizontal_display_end as u32 + 1) * clock_divisor * font_w as u32;
         let width = (self.crtc_horizontal_display_end as u32 + 1) * font_w as u32;
         let height = self.crtc_vertical_display_end as u32 + 1;
-        (width, height)
+        (width, height)*/
+        (320, 200)
+    }
+
+    /// Return the 16-bit value computed from the CRTC's pair of Page Address registers.
+    fn get_start_address(&self) -> u16 {
+        return self.crtc.start_address();
     }
 
     /// Unimplemented for indirect rendering.
@@ -146,7 +160,7 @@ impl VideoCard for EGACard {
     }
 
     fn get_clock_divisor(&self) -> u32 {
-        match self.sequencer_clocking_mode.dot_clock() {
+        match self.sequencer.clocking_mode.dot_clock() {
             DotClock::Native => 1,
             DotClock::HalfClock => 2,
         }
@@ -166,29 +180,25 @@ impl VideoCard for EGACard {
         self.mode_graphics
     }
 
-    /// Return the 16-bit value computed from the CRTC's pair of Page Address registers.
-    fn get_start_address(&self) -> u16 {
-        return (self.crtc_start_address_ho as u16) << 8 | self.crtc_start_address_lo as u16;
-    }
-
     fn get_cursor_info(&self) -> CursorInfo {
         let addr = self.get_cursor_address();
 
+        let span = self.crtc.get_cursor_span();
         match self.display_mode {
             DisplayMode::Mode0TextBw40 | DisplayMode::Mode1TextCo40 => CursorInfo {
                 addr: addr as usize,
                 pos_x: addr % 40,
                 pos_y: addr / 40,
-                line_start: self.crtc_cursor_start,
-                line_end: self.crtc_cursor_end,
+                line_start: span.0,
+                line_end: span.1,
                 visible: self.get_cursor_status(),
             },
             DisplayMode::Mode2TextBw80 | DisplayMode::Mode3TextCo80 => CursorInfo {
                 addr: addr as usize,
                 pos_x: addr % 80,
                 pos_y: addr / 80,
-                line_start: self.crtc_cursor_start,
-                line_end: self.crtc_cursor_end,
+                line_start: span.0,
+                line_end: span.1,
                 visible: self.get_cursor_status(),
             },
             _ => {
@@ -214,7 +224,7 @@ impl VideoCard for EGACard {
     }
 
     fn get_character_height(&self) -> u8 {
-        self.crtc_maximum_scanline + 1
+        self.crtc.maximum_scanline() + 1
     }
 
     /// Return the current palette number, intensity attribute bit, and alt color
@@ -269,47 +279,14 @@ impl VideoCard for EGACard {
         general_vec.push(("Adapter Type:".to_string(), VideoCardStateEntry::String(format!("{:?}", self.get_video_type()))));
         general_vec.push(("Display Mode:".to_string(), VideoCardStateEntry::String(format!("{:?}", self.get_display_mode()))));
         general_vec.push(("Pixel Clock:".to_string(), VideoCardStateEntry::String(format!("{:?}", self.misc_output_register.clock_select()))));
-        general_vec.push(("Clock Divisor:".to_string(), VideoCardStateEntry::String(format!("{:?}", self.clock_divisor))));
+        general_vec.push(("Clock Divisor:".to_string(), VideoCardStateEntry::String(format!("{:?}", self.sequencer.clock_divisor))));
         general_vec.push((
             "Field:".to_string(),
             VideoCardStateEntry::String(format!("{}x{}", self.extents.field_w, self.extents.field_h)),
         ));
 
         map.insert("General".to_string(), general_vec);
-
-        let mut crtc_vec = Vec::new();
-        push_reg_str!(crtc_vec, CRTCRegister::HorizontalTotal, "[R00]", self.crtc_horizontal_total);
-        push_reg_str!(crtc_vec, CRTCRegister::HorizontalDisplayEnd, "[R01]", self.crtc_horizontal_display_end);
-        push_reg_str!(crtc_vec, CRTCRegister::StartHorizontalBlank, "[R02]", self.crtc_start_horizontal_blank);
-        push_reg_str!(crtc_vec, CRTCRegister::EndHorizontalBlank, "[R03]", self.crtc_end_horizontal_blank.end_horizontal_blank());
-        push_reg_str!(crtc_vec, CRTCRegister::EndHorizontalBlank, "[R03:des]", self.crtc_end_horizontal_blank.display_enable_skew());
-        push_reg_str!(crtc_vec, CRTCRegister::EndHorizontalBlank, "[R03:norm]", self.crtc_end_horizontal_blank_norm);
-        push_reg_str!(crtc_vec, CRTCRegister::StartHorizontalRetrace, "[R04]", self.crtc_start_horizontal_retrace);
-        push_reg_str!(crtc_vec, CRTCRegister::EndHorizontalRetrace, "[R05]", self.crtc_end_horizontal_retrace.end_horizontal_retrace());
-        push_reg_str!(crtc_vec, CRTCRegister::EndHorizontalRetrace, "[R05:hrd]", self.crtc_end_horizontal_retrace.horizontal_retrace_delay());
-        push_reg_str!(crtc_vec, CRTCRegister::EndHorizontalRetrace, "[R05:som]", self.crtc_end_horizontal_retrace.start_odd());
-        push_reg_str!(crtc_vec, CRTCRegister::EndHorizontalRetrace, "[R05:norm]", self.crtc_end_horizontal_retrace_norm);
-        push_reg_str!(crtc_vec, CRTCRegister::VerticalTotal, "[R06]", self.crtc_vertical_total);
-        push_reg_str!(crtc_vec, CRTCRegister::Overflow, "[R07]", self.crtc_overflow);
-        push_reg_str!(crtc_vec, CRTCRegister::PresetRowScan, "[R08]", self.crtc_preset_row_scan);
-        push_reg_str!(crtc_vec, CRTCRegister::MaximumScanLine, "[R09]", self.crtc_maximum_scanline);
-        push_reg_str!(crtc_vec, CRTCRegister::CursorStartLine, "[R0A]", self.crtc_cursor_start);
-        push_reg_str!(crtc_vec, CRTCRegister::CursorEndLine, "[R0B]", self.crtc_cursor_end);
-        push_reg_str!(crtc_vec, CRTCRegister::StartAddressH, "[R0C]", self.crtc_start_address_ho);
-        push_reg_str!(crtc_vec, CRTCRegister::StartAddressL, "[R0D]", self.crtc_start_address_lo);
-        push_reg_str!(crtc_vec, CRTCRegister::CursorAddressH, "[R0E]", self.crtc_cursor_address_ho);
-        push_reg_str!(crtc_vec, CRTCRegister::CursorAddressL, "[R0F]", self.crtc_cursor_address_lo);
-        push_reg_str!(crtc_vec, CRTCRegister::VerticalRetraceStart, "[R10]", self.crtc_vertical_retrace_start);
-        push_reg_str!(crtc_vec, CRTCRegister::VerticalRetraceEnd, "[R11]", self.crtc_vertical_retrace_end.vertical_retrace_end());
-        push_reg_str!(crtc_vec, CRTCRegister::VerticalRetraceEnd, "[R11:norm]", self.crtc_vertical_retrace_end_norm);
-        push_reg_str!(crtc_vec, CRTCRegister::VerticalDisplayEnd, "[R12]", self.crtc_vertical_display_end);
-        push_reg_str!(crtc_vec, CRTCRegister::Offset, "[R13]", self.crtc_offset);
-        push_reg_str!(crtc_vec, CRTCRegister::UnderlineLocation, "[R14]", self.crtc_underline_location);
-        push_reg_str!(crtc_vec, CRTCRegister::StartVerticalBlank, "[R15]", self.crtc_start_vertical_blank);
-        push_reg_str!(crtc_vec, CRTCRegister::EndVerticalBlank, "[R16]", self.crtc_end_vertical_blank);
-        push_reg_str!(crtc_vec, CRTCRegister::ModeControl, "[R17]", self.crtc_mode_control);
-        push_reg_str!(crtc_vec, CRTCRegister::LineCompare, "[R18]", self.crtc_line_compare);
-        map.insert("CRTC".to_string(), crtc_vec);
+        map.insert("CRTC".to_string(), self.crtc.get_state());
 
         let mut external_vec = Vec::new();
         external_vec.push(("Misc Output".to_string(), VideoCardStateEntry::String(format!("{:?}", self.misc_output_register.into_bytes()[0]))));
@@ -323,109 +300,59 @@ impl VideoCard for EGACard {
         map.insert("External".to_string(), external_vec);
 
         let mut sequencer_vec = Vec::new();
-        sequencer_vec.push((format!("{:?}", SequencerRegister::Reset), VideoCardStateEntry::String(format!("{:02b}", self.sequencer_reset))));
-        sequencer_vec.push((format!("{:?}", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{:08b}", self.sequencer_clocking_mode.into_bytes()[0]))));           
-        sequencer_vec.push((format!("{:?} [cc]", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{:?}", self.sequencer_clocking_mode.character_clock()))));
-        sequencer_vec.push((format!("{:?} [bw]", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{}", self.sequencer_clocking_mode.bandwidth()))));
-        sequencer_vec.push((format!("{:?} [sl]", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{}", self.sequencer_clocking_mode.shift_load()))));
-        sequencer_vec.push((format!("{:?} [dc]", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{:?}", self.sequencer_clocking_mode.dot_clock()))));
+        sequencer_vec.push((format!("{:?}", SequencerRegister::Reset), VideoCardStateEntry::String(format!("{:02b}", self.sequencer.reset))));
+        sequencer_vec.push((format!("{:?}", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{:04b}", self.sequencer.clocking_mode.into_bytes()[0]))));           
+        sequencer_vec.push((format!("{:?} [cc]", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{:?}", self.sequencer.clocking_mode.character_clock()))));
+        sequencer_vec.push((format!("{:?} [bw]", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{}", self.sequencer.clocking_mode.bandwidth()))));
+        sequencer_vec.push((format!("{:?} [sl]", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{}", self.sequencer.clocking_mode.shift_load()))));
+        sequencer_vec.push((format!("{:?} [dc]", SequencerRegister::ClockingMode), VideoCardStateEntry::String(format!("{:?}", self.sequencer.clocking_mode.dot_clock()))));
 
-        sequencer_vec.push((format!("{:?}", SequencerRegister::MapMask), VideoCardStateEntry::String(format!("{:04b}", self.sequencer_map_mask))));
-        sequencer_vec.push((format!("{:?}", SequencerRegister::CharacterMapSelect), VideoCardStateEntry::String(format!("{}", self.sequencer_character_map_select))));
-        sequencer_vec.push((format!("{:?}", SequencerRegister::MemoryMode), VideoCardStateEntry::String(format!("{}", self.sequencer_memory_mode))));
+        sequencer_vec.push((format!("{:?}", SequencerRegister::MapMask), VideoCardStateEntry::String(format!("{:04b}", self.sequencer.map_mask))));
+        sequencer_vec.push((format!("{:?}", SequencerRegister::CharacterMapSelect), VideoCardStateEntry::String(format!("{:04b}", self.sequencer.character_map_select))));
+        sequencer_vec.push((format!("{:?}", SequencerRegister::MemoryMode), VideoCardStateEntry::String(format!("{:04b}", self.sequencer.memory_mode.into_bytes()[0]))));
+        sequencer_vec.push((format!("{:?} [ag]", SequencerRegister::MemoryMode), VideoCardStateEntry::String(format!("{}", self.sequencer.memory_mode.alpha_mode() as u8))));
+        sequencer_vec.push((format!("{:?} [em]", SequencerRegister::MemoryMode), VideoCardStateEntry::String(format!("{}", self.sequencer.memory_mode.extended_memory() as u8))));
+        sequencer_vec.push((format!("{:?} [oe]", SequencerRegister::MemoryMode), VideoCardStateEntry::String(format!("{}", self.sequencer.memory_mode.odd_even()))));
         map.insert("Sequencer".to_string(), sequencer_vec);
 
-        let mut graphics_vec = Vec::new();
-        graphics_vec.push((format!("{:?}", GraphicsRegister::SetReset), VideoCardStateEntry::String(format!("{:04b}", self.graphics_set_reset))));
-        graphics_vec.push((format!("{:?}", GraphicsRegister::EnableSetReset), VideoCardStateEntry::String(format!("{:04b}", self.graphics_enable_set_reset))));
-        graphics_vec.push((format!("{:?}", GraphicsRegister::ColorCompare), VideoCardStateEntry::String(format!("{:04b}", self.graphics_color_compare))));
-        graphics_vec.push((format!("{:?} [fn]", GraphicsRegister::DataRotate), VideoCardStateEntry::String(format!("{:?}", self.graphics_data_rotate.function()))));
-        graphics_vec.push((format!("{:?} [ct]", GraphicsRegister::DataRotate), VideoCardStateEntry::String(format!("{:?}", self.graphics_data_rotate.count()))));              
-        graphics_vec.push((format!("{:?}", GraphicsRegister::ReadMapSelect), VideoCardStateEntry::String(format!("{:03b}", self.graphics_read_map_select))));
-
-        graphics_vec.push((format!("{:?} [sr]", GraphicsRegister::Mode), VideoCardStateEntry::String(format!("{:?}", self.graphics_mode.shift_mode()))));
-        graphics_vec.push((format!("{:?} [o/e]", GraphicsRegister::Mode), VideoCardStateEntry::String(format!("{:?}", self.graphics_mode.odd_even()))));
-        graphics_vec.push((format!("{:?} [rm]", GraphicsRegister::Mode), VideoCardStateEntry::String(format!("{:?}",self.graphics_mode.read_mode()))));
-        graphics_vec.push((format!("{:?} [tc]", GraphicsRegister::Mode), VideoCardStateEntry::String(format!("{:?}", self.graphics_mode.test_condition()))));
-        graphics_vec.push((format!("{:?} [wm]", GraphicsRegister::Mode), VideoCardStateEntry::String(format!("{:?}", self.graphics_mode.write_mode()))));
-
-        graphics_vec.push((format!("{:?} [gm]", GraphicsRegister::Miscellaneous), VideoCardStateEntry::String(format!("{:?}", self.graphics_micellaneous.graphics_mode()))));
-        graphics_vec.push((format!("{:?} [coe]", GraphicsRegister::Miscellaneous), VideoCardStateEntry::String(format!("{:?}", self.graphics_micellaneous.chain_odd_even()))));
-        graphics_vec.push((format!("{:?} [mm]", GraphicsRegister::Miscellaneous), VideoCardStateEntry::String(format!("{:?}", self.graphics_micellaneous.memory_map()))));            
-
-        graphics_vec.push((format!("{:?}", GraphicsRegister::ColorDontCare), VideoCardStateEntry::String(format!("{:04b}", self.graphics_color_dont_care))));
-        graphics_vec.push((format!("{:?}", GraphicsRegister::BitMask), VideoCardStateEntry::String(format!("{:08b}", self.graphics_bitmask))));
+        let graphics_vec = self.gc.get_state();
         map.insert("Graphics".to_string(), graphics_vec);
-
-        /* old-style attribute palette
-        let mut attribute_pal_vec = Vec::new();
-        for i in 0..16 {
-            attribute_pal_vec.push((format!("Palette register {}", i),
-                VideoCardStateEntry::String(format!("{:06b}", self.attribute_palette_registers[i]))
-            ));
-        }
-        map.insert("AttributePalette".to_string(), attribute_pal_vec);
-        */
 
         let mut attribute_pal_vec = Vec::new();
         for i in 0..16 {
             // Attribute palette entries are interpreted differently depending on the current clock speed
             // Low resolution modes use 4BPP palette entries, high resolution modes use 6bpp.
             let pal_resolved = match self.misc_output_register.clock_select() {
-                ClockSelect::Clock14 => self.attribute_palette_registers[i].four_to_six,
-                _ => self.attribute_palette_registers[i].six,
+                ClockSelect::Clock14 => self.ac.attribute_palette_registers[i].four_to_six,
+                _ => self.ac.attribute_palette_registers[i].six,
             };
 
             let (r, g, b) = EGACard::ega_to_rgb(pal_resolved);
             attribute_pal_vec.push((
                 format!("{}", i),
-                VideoCardStateEntry::Color(format!("{:06b}", self.attribute_palette_registers[i].six), r, g, b),
+                VideoCardStateEntry::Color(format!("{:06b}", self.ac.attribute_palette_registers[i].six), r, g, b),
             ));
         }
         map.insert("AttributePalette".to_string(), attribute_pal_vec);
-
-        let mut attribute_vec = Vec::new();
-        attribute_vec.push((format!("{:?} mode:", AttributeRegister::ModeControl), VideoCardStateEntry::String(format!("{:?}", self.attribute_mode_control.mode()))));
-        attribute_vec.push((format!("{:?} disp:", AttributeRegister::ModeControl), VideoCardStateEntry::String(format!("{:?}", self.attribute_mode_control.display_type()))));
-        attribute_vec.push((format!("{:?} elgc:", AttributeRegister::ModeControl), VideoCardStateEntry::String(format!("{:?}", self.attribute_mode_control.enable_line_character_codes()))));
-        attribute_vec.push((format!("{:?} attr:", AttributeRegister::ModeControl), VideoCardStateEntry::String(format!("{:?}", self.attribute_mode_control.enable_blink_or_intensity()))));            
-
-        let (r, g, b) = EGACard::ega_to_rgb( self.attribute_overscan_color.into_bytes()[0]);
-        attribute_vec.push((format!("{:?}", AttributeRegister::OverscanColor), VideoCardStateEntry::Color(format!("{:06b}", self.attribute_overscan_color.into_bytes()[0]), r, g, b)));
-        attribute_vec.push((format!("{:?} en:", AttributeRegister::ColorPlaneEnable), VideoCardStateEntry::String(format!("{:04b}", self.attribute_color_plane_enable.enable_plane()))));           
-        attribute_vec.push((format!("{:?} mux:", AttributeRegister::ColorPlaneEnable), VideoCardStateEntry::String(format!("{:02b}", self.attribute_color_plane_enable.video_status_mux()))));                
-        attribute_vec.push((format!("{:?}", AttributeRegister::HorizontalPelPanning), VideoCardStateEntry::String(format!("{}", self.attribute_pel_panning))));     
-        //attribute_overscan_color: AOverscanColor::new(),
-        //attribute_color_plane_enable: AColorPlaneEnable::new(),
-        map.insert("Attribute".to_string(), attribute_vec);
+        map.insert("Attribute".to_string(), self.ac.get_state());
+        map.insert("Crtc Counters".to_string(), self.crtc.get_counter_state());
 
         let mut internal_vec = Vec::new();
+        internal_vec.push(("scanline:".to_string(), VideoCardStateEntry::String(format!("{}", self.scanline))));
+        internal_vec.push(("vma:".to_string(), VideoCardStateEntry::String(format!("{:04X}", self.vma))));
+        internal_vec.push(("rba:".to_string(), VideoCardStateEntry::String(format!("{:04X}", self.rba))));
+        internal_vec.push(("raster_x:".to_string(), VideoCardStateEntry::String(format!("{}", self.raster_x))));
+        internal_vec.push(("raster_y:".to_string(), VideoCardStateEntry::String(format!("{}", self.raster_y))));
 
-        internal_vec.push((format!("hcc:"), VideoCardStateEntry::String(format!("{}", self.hcc))));
-        internal_vec.push((format!("vlc:"), VideoCardStateEntry::String(format!("{}", self.vlc))));
-        internal_vec.push((format!("vcc:"), VideoCardStateEntry::String(format!("{}", self.vcc))));
-        internal_vec.push((format!("hslc:"), VideoCardStateEntry::String(format!("{}", self.hslc))));
-        internal_vec.push((format!("scanline:"), VideoCardStateEntry::String(format!("{}", self.scanline))));
-        internal_vec.push((format!("hsc:"), VideoCardStateEntry::String(format!("{}", self.hsc))));
-        internal_vec.push((format!("vma:"), VideoCardStateEntry::String(format!("{:04X}", self.vma))));
-        internal_vec.push((format!("vma':"), VideoCardStateEntry::String(format!("{:04X}", self.vma_t))));
-        internal_vec.push((format!("vmws:"), VideoCardStateEntry::String(format!("{}", self.vmws))));
-        internal_vec.push((format!("rba:"), VideoCardStateEntry::String(format!("{:04X}", self.rba))));
-        internal_vec.push((format!("de:"), VideoCardStateEntry::String(format!("{}", self.in_display_area))));
-        internal_vec.push((format!("crtc_hblank:"), VideoCardStateEntry::String(format!("{}", self.crtc_hblank))));
-        internal_vec.push((format!("crtc_vblank:"), VideoCardStateEntry::String(format!("{}", self.crtc_vblank))));
-        internal_vec.push((format!("raster_x:"), VideoCardStateEntry::String(format!("{}", self.raster_x))));
-        internal_vec.push((format!("raster_y:"), VideoCardStateEntry::String(format!("{}", self.raster_y))));
-        internal_vec.push((format!("border:"), VideoCardStateEntry::String(format!("{}", self.crtc_hborder))));
         //internal_vec.push((format!("s_reads:"), VideoCardStateEntry::String(format!("{}", self.status_reads))));
         //internal_vec.push((format!("missed_hsyncs:"), VideoCardStateEntry::String(format!("{}", self.missed_hsyncs))));
         //internal_vec.push((format!("vsync_cycles:"), VideoCardStateEntry::String(format!("{}", self.cycles_per_vsync))));
         //internal_vec.push((format!("cur_screen_cycles:"), VideoCardStateEntry::String(format!("{}", self.cur_screen_cycles))));
         //internal_vec.push((format!("phase:"), VideoCardStateEntry::String(format!("{}", self.cycles & 0x0F))));
-        internal_vec.push((format!("blink state:"), VideoCardStateEntry::String(format!("{}", self.blink_state))));
 
-        internal_vec.push((format!("hsync_ct:"), VideoCardStateEntry::String(format!("{}", self.hsync_ct))));
-        internal_vec.push((format!("vsync_ct:"), VideoCardStateEntry::String(format!("{}", self.vsync_ct))));
+        internal_vec.push(("blink state:".to_string(), VideoCardStateEntry::String(format!("{}", self.blink_state))));
+        internal_vec.push(("hsync_ct:".to_string(), VideoCardStateEntry::String(format!("{}", self.hsync_ct))));
+        internal_vec.push(("vsync_ct:".to_string(), VideoCardStateEntry::String(format!("{}", self.vsync_ct))));
 
         map.insert("Internal".to_string(), internal_vec);
 
@@ -511,7 +438,7 @@ impl VideoCard for EGACard {
     }
 
     fn get_pixel_raw(&self, x: u32, y: u32) -> u8 {
-        let mut byte = 0;
+        /*        let mut byte = 0;
 
         let x_byte_offset = (x + self.attribute_pel_panning as u32) / 8;
         let x_bit_offset = (x + self.attribute_pel_panning as u32) % 8;
@@ -533,9 +460,9 @@ impl VideoCard for EGACard {
             read_offset = (y_offset + x_byte_offset + self.crtc_start_address as u32) as usize;
         }
 
-        if read_offset < self.planes[0].buf.len() {
+        if read_offset < self.sequencer.vram.plane_len() {
             for i in 0..4 {
-                let read_byte = self.planes[i].buf[read_offset];
+                let read_byte = self.sequencer.vram.read_u8(i, read_offset);
                 let read_bit = match read_byte & (0x01 << (7 - x_bit_offset)) != 0 {
                     true => 1,
                     false => 0,
@@ -546,12 +473,12 @@ impl VideoCard for EGACard {
             }
             // return self.attribute_palette_registers[byte & 0x0F].into_bytes()[0];
             return self.attribute_palette_registers[byte & 0x0F].six;
-        }
+        }*/
         0
     }
 
     fn get_plane_slice(&self, plane: usize) -> &[u8] {
-        &self.planes[plane].buf
+        self.sequencer.vram.plane_slice(plane)
     }
 
     fn dump_mem(&self, path: &Path) {
@@ -559,7 +486,7 @@ impl VideoCard for EGACard {
             let mut filename = path.to_path_buf();
             filename.push(format!("ega_plane{}.bin", i));
 
-            match std::fs::write(filename.clone(), &self.planes[i].buf) {
+            match std::fs::write(filename.clone(), &self.get_plane_slice(i)) {
                 Ok(_) => {
                     log::debug!("Wrote memory dump: {}", &filename.display())
                 }
